refactor(types): split shared/types/index.ts into domain-focused files
The 323-line barrel mixed generic primitives, AI/provider config, framework
context, the core finding shapes, pipeline identity, and per-stage metadata
for five different stages. Split by domain/purpose with speaking names:

  primitives.ts · ai-config.ts · framework-context.ts · finding.ts ·
  pipeline.ts · analysis-metadata.ts · review-metadata.ts ·
  root-cause-metadata.ts · fix-metadata.ts · report-metadata.ts ·
  judge-metadata.ts

index.ts is now a pure barrel re-exporting them, so every existing
`from '.../shared/types'` import keeps resolving — zero consumer churn.
Type definitions are copied verbatim (no field renames — those would be a
behavior change on serialized shapes). The drifted second FrameworkContext
in framework-detector.ts is noted for a separate reconciliation.

No behavior change: build clean, 2619/2619 tests pass, lint clean.
